alter csv file to allow for lat long, update schema accordingly.

+++ db/migrate/20171005211223_add_lat_long_to_stations.rb
@@ -1,6 +1,6 @@
 class AddLatLongToStations < ActiveRecord::Migration[5.1]
   def change
-    add_column :stations, :lat, :integer
-    add_column :stations, :lng, :integer
+    add_column :stations, :lat, :float
+    add_column :stations, :lng, :float
   end
 end
